IMDb API
This uses the IMDb API and uses the URL of a specific movie to output a JSON structure.  Then, it will convert to python structure that can be referenced by the fields. You can call the title, rating, etc and it will output. 

Later, you can loop through it and put multiple movies into the text files.